+++ core/src/com/raccoon/mygame/obstacle/BoxObstacle.java
@@ -15,7 +15,7 @@ import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.physics.box2d.Fixture;
 import com.badlogic.gdx.physics.box2d.PolygonShape;
-import com.raccoon.mygame.game.GameCanvas;
+import com.raccoon.mygame.view.GameCanvas;
 
 /**
  * Box-shaped model to support collisions.
