+++ core/src/com/raccoon/mygame/obstacle/SimpleObstacle.java
@@ -24,7 +24,7 @@ import com.badlogic.gdx.physics.box2d.BodyDef.BodyType;
 import com.badlogic.gdx.physics.box2d.Filter;
 import com.badlogic.gdx.physics.box2d.Fixture;
 import com.badlogic.gdx.physics.box2d.World;
-import com.raccoon.mygame.game.GameCanvas;
+import com.raccoon.mygame.view.GameCanvas;
 
 /**
  * Base model class to support collisions.
