+++ core/src/com/raccoon/mygame/view/GameCanvas.java
@@ -15,7 +15,7 @@
  * Based on original PhysicsDemo Lab by Don Holden, 2007
  * LibGDX version, 2/6/2015
  */
-package com.raccoon.mygame.game;
+package com.raccoon.mygame.view;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
